implementação de login e senha hash para administrador, e autenticação de rotas privadas de administrador

+++ src/app/api/categorias/[id]/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { CategoriasService } from "@/modules/categorias/categoriasService";
 import { AppError } from "@/shared/errors/AppError";
+import { authAdmin } from "@/middlewares/authAdminMiddleware";
 
 export async function GET(req: NextRequest, context: { params: Promise<{ id: string }> }) {
   try {
@@ -18,6 +19,10 @@ export async function GET(req: NextRequest, context: { params: Promise<{ id: str
 
 export async function DELETE(req: NextRequest, context: { params: Promise<{ id: string }> }) {
   try {
+    const auth = authAdmin(req)
+    if (!auth) {
+      return NextResponse.json({ message: "Token inválido" }, { status: 401 })
+    }
     const { id } = await context.params
     const categoria = await CategoriasService.deleteCategorias(id)
     return NextResponse.json({ data: categoria, message: "Categoria deletada com sucesso" }, { status: 200 })
